extract tracer to middleware
