+++ app.js
@@ -1,14 +1,16 @@
-var express = require('express');
-var logger = require('morgan');
+const express = require('express');
+const tracer = require('./middleware/tracer');
+const logger = require('morgan');
 const bodyParser = require('body-parser');
-var cookieParser = require('cookie-parser');
-var path = require('path');
+const cookieParser = require('cookie-parser');
+const path = require('path');
 
-var indexRouter = require('./routes/index');
-var usersRouter = require('./routes/users');
+const indexRouter = require('./routes/index');
+const usersRouter = require('./routes/users');
 
-var app = express();
+const app = express();
 
+app.use(tracer());
 app.use(logger('dev'));
 // app.use(express.json());
 // app.use(express.urlencoded({ extended: false }));
